requirements.txt
Django==3.1.5
django-filebrowser==3.13.2
django-tinymce==3.3.0
